requirements.txt

pandas>=2.2
numpy>=1.26
pyyaml>=6.0
requests>=2.32
matplotlib>=3.8
statsmodels>=0.14
scikit-learn>=1.5
xgboost>=2.0
typer>=0.12
pydantic>=2.8
tqdm>=4.66
yfinance>=0.2
python-dateutil>=2.9
joblib>=1.4

streamlit>=1.38
